Fix de crash quand l'api etait utilisee

+++ app/controllers/history_controller.rb
@@ -14,14 +14,33 @@ class HistoryController < ActionController::API
   end
 
   def create
-    history = History.create(email: 'totot@gmail.com', url: 'www.fakeurl.fake', title: 'My Title')
+    enter_date = params[:enter_date]
+    leave_date = params[:leave_date]
+    focus_elasped_time = params[:focus_elasped_time]
+    total_elasped_time = params[:total_elasped_time]
+    history = History.create(email: params[:email],
+                             url: params[:url],
+                             origin_url: params[:origin_url],
+                             target_url: params[:target_url],
+                             lang: params[:lang],
+                             title: params[:title],
+                             description: params[:description],
+                             image: params[:image],
+                             search: params[:search],
+                             enter_date: enter_date,
+                             leave_date: leave_date,
+                             focus_elasped_time: focus_elasped_time,
+                             total_elasped_time: total_elasped_time)
     history.save
+    render json: history
   end
 
   def destroy
     id = params[:id]
     history = History.find_by(id: id)
-    history.destroy
-    history.save
+    if history != nil
+      history.destroy
+    end
+    render json: history
   end
 end
